+++ src/Dracodeum/Kit/Interfaces/DebugInfo.php
@@ -18,10 +18,8 @@ interface DebugInfo
 	/**
 	 * Get debug info.
 	 * 
-	 * @param bool $recursive [default = false]
-	 * <p>Get debug info from all the possible referenced subobjects recursively (if applicable).</p>
 	 * @return array
 	 * <p>The debug info.</p>
 	 */
-	public function getDebugInfo(bool $recursive = false): array;
+	public function getDebugInfo(): array;
 }